Refactor "Subscribe in Calendar" to not use a modal
- Fundamentally against modals; it's not that big of a deal to just show it on the page here
- Now we just have CalendarUrlCopy as a component for displaying the URL and handling the clipboard stuff
- Also updated the headings and spacing on the Upcoming page a big, using more space between sections and using SectionHeading instead of PageTitle for each part
